updateConfig

+++ updateConfigs.sh
@@ -1,5 +1,15 @@
 #!/bin/bash
 
+
+die () {
+    echo >&2 "$@"
+    exit 1
+}
+
+
+[ "$#" -eq 1 ] || die "1 argument required, $# provided"
+
+
 rm -rf ./qtile && echo "removed qtile"
 rm -rf ./picom && echo "removed picom"
 rm -rf ./autokey && echo "removed autokey"
@@ -19,6 +29,6 @@ mkdir bash && echo "Created bash directory"
 cp -r ~/.bashrc ./bash && echo "Copied bashrc configs"
 cp -r ~/.shell_aliases ./bash && echo "Copied shell_aliases configs"
 
-git add --all
-git commit -m $1
+git add --all && \
+git commit -m $1 && \
 git push
